+++ src/vldp/vldp.cpp
@@ -45,11 +45,11 @@ unsigned int g_ack_count = ACK_COUNT_INITIAL; // the result returned by the
                                               // internal child thread
 char g_req_file[STRSIZE];                     // requested mpeg filename
 Uint32 g_req_timer       = 0; // requests timer value to be used for mpeg playback
-Uint16 g_req_frame       = 0; // requested frame to search to
+Uint32 g_req_frame       = 0; // requested frame to search to
 Uint32 g_req_min_seek_ms = 0; // seek must take at least this many milliseconds
                               // (simulate laserdisc seek delay)
 VLDP_BOOL g_req_precache          = VLDP_FALSE; // whether g_req_idx has any meaning
-unsigned int g_req_idx            = 0; // multipurpose index (used by precaching)
+unsigned long g_req_idx            = 0; // multipurpose index (used by precaching)
 unsigned int g_req_skip_per_frame = 0; // how many frames to skip per frame (for
                                        // playing at 2X for example)
 unsigned int g_req_stall_per_frame = 0; // how many frames to stall per frame
@@ -173,7 +173,7 @@ int vldp_open(const char *filename)
     return result;
 }
 
-VLDP_BOOL vldp_open_precached(unsigned int uIdx, const char *filename)
+VLDP_BOOL vldp_open_precached(unsigned long uIdx, const char *filename)
 {
     VLDP_BOOL bResult = VLDP_FALSE;
 
@@ -224,7 +224,7 @@ VLDP_BOOL vldp_precache(const char *filename)
 
 // issues search command and returns immediately to parent thread.
 // Search will not be complete until the VLDP status is STAT_PAUSED
-int vldp_search(Uint16 frame, Uint32 min_seek_ms)
+int vldp_search(Uint32 frame, Uint32 min_seek_ms)
 {
     int result = 0;
 
@@ -237,7 +237,7 @@ int vldp_search(Uint16 frame, Uint32 min_seek_ms)
 }
 
 // issues search command blocks until search is complete
-int vldp_search_and_block(Uint16 frame, Uint32 min_seek_ms)
+int vldp_search_and_block(Uint32 frame, Uint32 min_seek_ms)
 {
     int result = 0;
 
@@ -263,7 +263,7 @@ int vldp_play(Uint32 timer)
     return (result);
 }
 
-int vldp_skip(Uint16 frame)
+int vldp_skip(Uint32 frame)
 {
     int result = 0;
